ppc4xx: Remove compiler warning from previous commit

+++ post/cpu/ppc4xx/cache.c
@@ -119,7 +119,7 @@ int cache_post_test (int flags)
 	if (ints)
 		enable_interrupts ();
 
-	remove_tlb(virt, CACHE_POST_SIZE);
+	remove_tlb((u32)virt, CACHE_POST_SIZE);
 
 	return res;
 }
